tools/system-design-space-importer/src/system_design_space_importer/yamlio.py
from pathlib import Path

import yaml


def write_yaml(path, payload):
    target = Path(path)
    target.parent.mkdir(parents=True, exist_ok=True)
    with target.open("w", encoding="utf-8") as handle:
        yaml.safe_dump(payload, handle, allow_unicode=True, sort_keys=False)
