tools/system-design-space-importer/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "system-design-space-importer"
version = "0.1.0"
description = "Importer tool for system-design.space draft topic packages"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
sds-importer = "system_design_space_importer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
